remove _id, id and __v; modify the response structure for airport controllers
