+++ models/airports/GNS430_model/updateGns430AirportModel.js
@@ -116,6 +116,8 @@ mongoose.plugin((gns430AirportSchema) => {
         transform(doc, ret) {
             ret.id = ret._id;
             delete ret._id;
+            delete ret.__v;
+            delete ret.id;
         }
     };
 });
